# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ToolRoute API",
    "description": "The OpenRouter for Tools — unified API gateway for 50+ AI agent tools. One API key, every tool.",
    "version": "1.0.0",
    "contact": {
      "url": "https://toolroute.ai",
      "email": "support@toolroute.ai"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "https://toolroute.ai",
      "description": "Production"
    }
  ],
  "security": [],
  "paths": {
    "/api/v1/execute": {
      "post": {
        "operationId": "executeTool",
        "summary": "Execute a tool",
        "description": "Execute any tool in the ToolRoute registry via the unified gateway. Requires an API key (tr_live_xxx). The gateway routes to the best available provider, handles auth, and returns structured output.",
        "tags": ["Gateway"],
        "security": [{ "apiKey": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ExecuteRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tool executed successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ExecuteResponse" }
              }
            }
          },
          "400": {
            "description": "Invalid request (missing tool, bad input)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "402": {
            "description": "Insufficient balance",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "404": {
            "description": "Tool not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "500": {
            "description": "Internal server error or upstream provider failure",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/key": {
      "get": {
        "operationId": "getApiKeyInfo",
        "summary": "Get API key info and balance",
        "description": "Returns information about the authenticated API key including current balance, rate limits, and usage summary.",
        "tags": ["Keys"],
        "security": [{ "apiKey": [] }],
        "responses": {
          "200": {
            "description": "API key information",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiKeyInfo" }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/checkout": {
      "post": {
        "operationId": "createCheckoutSession",
        "summary": "Create Stripe checkout session",
        "description": "Creates a Stripe checkout session to add credits to the user's account. Redirects to Stripe for payment.",
        "tags": ["Billing"],
        "security": [{ "sessionAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["amount"],
                "properties": {
                  "amount": {
                    "type": "number",
                    "description": "Amount in USD to add as credits",
                    "minimum": 5,
                    "example": 10
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Checkout session created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "url": {
                      "type": "string",
                      "format": "uri",
                      "description": "Stripe checkout URL to redirect the user to"
                    },
                    "session_id": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key in the format tr_live_xxx. Pass as Bearer token in Authorization header."
      },
      "sessionAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Supabase JWT session token from dashboard login. Pass as Bearer token in Authorization header."
      }
    },
    "schemas": {
      "ExecuteRequest": {
        "type": "object",
        "required": ["tool", "input"],
        "properties": {
          "tool": {
            "type": "string",
            "description": "Tool identifier in provider/operation format",
            "example": "resend/send_email"
          },
          "input": {
            "type": "object",
            "description": "Tool-specific input parameters",
            "additionalProperties": true,
            "example": {
              "to": "user@example.com",
              "subject": "Hello",
              "body": "World"
            }
          },
          "provider": {
            "type": "object",
            "description": "Optional provider routing preferences",
            "properties": {
              "prefer": {
                "type": "string",
                "description": "Preferred provider name"
              },
              "allow_fallbacks": {
                "type": "boolean",
                "default": true,
                "description": "Whether to fall back to alternative providers on failure"
              }
            }
          }
        }
      },
      "ExecuteResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique request ID for this execution",
            "example": "req_abc123"
          },
          "tool": {
            "type": "string",
            "description": "Tool that was executed",
            "example": "resend/send_email"
          },
          "provider": {
            "type": "string",
            "description": "Provider that handled the request",
            "example": "resend"
          },
          "data": {
            "type": "object",
            "description": "Tool output data (structure varies by tool)",
            "additionalProperties": true
          },
          "usage": {
            "$ref": "#/components/schemas/UsageInfo"
          }
        }
      },
      "UsageInfo": {
        "type": "object",
        "properties": {
          "cost": {
            "type": "number",
            "description": "Cost of this request in USD",
            "example": 0.001
          },
          "balance_remaining": {
            "type": "number",
            "description": "Remaining account balance in USD",
            "example": 4.99
          },
          "latency_ms": {
            "type": "integer",
            "description": "End-to-end latency in milliseconds",
            "example": 234
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string",
                "description": "Human-readable error message",
                "example": "Insufficient balance. Add credits at https://toolroute.ai/dashboard/billing"
              },
              "code": {
                "type": "string",
                "description": "Machine-readable error code",
                "enum": [
                  "invalid_request",
                  "authentication_required",
                  "invalid_api_key",
                  "insufficient_balance",
                  "tool_not_found",
                  "rate_limit_exceeded",
                  "provider_error",
                  "internal_error"
                ],
                "example": "insufficient_balance"
              }
            }
          }
        }
      },
...
      "ApiKeyInfo": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "key_hint": { "type": "string", "example": "tr_live_...f456" },
          "balance": {
            "type": "number",
            "description": "Current account balance in USD",
            "example": 9.85
          },
          "rate_limit": {
            "type": "object",
            "properties": {
              "requests_per_minute": { "type": "integer", "example": 60 },
              "requests_remaining": { "type": "integer", "example": 58 }
            }
          },
          "created_at": { "type": "string", "format": "date-time" },
          "last_used_at": { "type": "string", "format": "date-time", "nullable": true },
          "total_requests": { "type": "integer" },
          "total_spend": { "type": "number", "description": "Total amount spent in USD" }
        }
      },
...
    }
  },
...
}